dma-buf: heaps: carveout: Add DMA sync ops for cached buffers
When originally adding support for cached carveout buffers we made some
assumptions about the devices this would run on, most important here is
that devices with buffer attachable DMA accelerators would all be K3
devices with MSMC coherency support.

With the introduction of AM62a this does not hold true. Lets do this
the right way here and add the needed sync operation handlers in the
caching path.

Fixes: 5b692a0c8157 ("HACK: dma-buf: heaps: carveout: Add support for cached carveout heaps")
Signed-off-by: Andrew Davis <afd@ti.com>