HACK: dt-bindings: misc: Add ti,dma-buf-phys binding doc
Driver allowing user-space attaching of DMA-BUFs returning CPU physical
addresses. The reasoning for making this act like a regular device
described by DT is so this virtual device that binds the buffer can be
made to act as if it is out on a bus segment or behind an IOMMU.

Add this binding here.

Signed-off-by: Andrew Davis <afd@ti.com>