ti_config_fragments: baseport: Trim out defaults
All carried changes to the upstream kernel that are needed to boot it on
TI platforms should be considered tech debt. Changes to the default ARMv8
configuration are no different. We need to reduced these over time by
working to get these options in the upstream defconfig. To start to see
where we stand today, lets remove all the options that do not change the
upstream default. Then anything remaining in this file should be
investigated for appropriateness in upstream and sent accordingly.

Signed-off-by: Andrew Davis <afd@ti.com>