spi: cadence-quadspi: Implement Suspend Resume
The cqspi suspend and resume functions were earlier not working
correctly when attempting to suspend the device. Code was tested on a TI
AM62x SK EVM, and prior to this fix the suspend as well as resume used
to fail in this qspi driver. This was due to the fact that there was no
proper code to restore the necessary registers context like the phy
config regs and delay reg. This caused the data read from the flash
after the resume to be corrupted and unusable.

The fix has now been verified after comparing data written to the flash
and reading it back, both before and after resume and made sure it all
matches.

Signed-off-by: Dhruva Gole <d-gole@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>