ti_config_fragments: enable CONFIG_DEBUG_INFO
This allows proper source-and-assembly debugging using JTAG and also
allows us to debug WARN_ONs and other backtraces without having to
recompile the kernel (which could change the system behavior).

We already do this for u-boot/SPL, so there's no strong reason not to do
this in the kernel as well.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Acked-by: Andrew Davis <afd@ti.com>